Add FUZZING support
This adds the FUZZING variable to the build system, which will build all
code with libFuzzer instrumentation and activate the pfuzz subdirectory.
Inside this subdirectory, all .c files will be built as standalone
fuzzing targets. This flag is supposed to be combined with SANITIZE=1
but there are cases where it is desirable to build without sanitizers,
hence I left the flags separated.